+++ src/app/modules/map/service/map.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@angular/core';
 import { SnackBarService } from '@shared/module/material/service';
 import { ItemClipboardResultCode, ItemClipboardService } from '@shared/module/poe/service';
 import { ItemCategory, ItemSection } from '@shared/module/poe/type';
-import { Observable } from 'rxjs';
+import { Observable, throwError } from 'rxjs';
+import { catchError, flatMap } from 'rxjs/operators';
 import { MapUserSettings } from '../component/map-settings/map-settings.component';
 import { MapDialogService } from './map-dialog.service';
 
@@ -17,24 +18,30 @@ export class MapService {
     }
 
     public info(settings: MapUserSettings): Observable<void> {
-        const result = this.itemClipboard.copy({
+        return this.itemClipboard.copy({
             [ItemSection.Rartiy]: true,
             [ItemSection.ItemLevel]: true,
             [ItemSection.Properties]: true,
             [ItemSection.Stats]: true,
-        });
-        switch (result.code) {
-            case ItemClipboardResultCode.Success:
-                if (result.item.category !== ItemCategory.Map) {
-                    return this.snackbar.warning('Item was not a map.');
+        }).pipe(
+            flatMap(result => {
+                switch (result.code) {
+                    case ItemClipboardResultCode.Success:
+                        if (result.item.category !== ItemCategory.Map) {
+                            return this.snackbar.warning('Item was not a map.');
+                        }
+                        return this.dialogService.open(result.point, result.item, settings);
+                    case ItemClipboardResultCode.Empty:
+                        return this.snackbar.warning('Clipboard text was empty. Make sure the game is focused.');
+                    case ItemClipboardResultCode.ParserError:
+                        return this.snackbar.warning('Copied item could not be parsed. Make sure you have the correct language selected.');
+                    default:
+                        return throwError(`Code: '${result.code}' out of range`);
                 }
-                return this.dialogService.open(result.point, result.item, settings);
-            case ItemClipboardResultCode.Empty:
-                return this.snackbar.warning('Clipboard text was empty. Make sure the game is focused.');
-            case ItemClipboardResultCode.ParserError:
-                return this.snackbar.warning('Copied item could not be parsed. Make sure you have the correct language selected.');
-            case ItemClipboardResultCode.Error:
+            }),
+            catchError(() => {
                 return this.snackbar.error('An unexpected error occured while parsing the item.');
-        }
+            })
+        );
     }
 }